Revert "Rename failed_commnad_error.py to exceptions.py."
This reverts commit d6aed1c91767e510961d57d2736d93d9a374c823.
