+++ src/pre_commit_sbt/err/failed_command_error.py
