+++ tests/integration/run_sbt_command_integration_test.py
@@ -6,7 +6,7 @@ from testing_utils.sbt import add_sleep_command_to_sbt
 from testing_utils.sbt import add_touch_command_to_sbt
 
 from pre_commit_sbt.err.error_msgs import COMMAND_FAILED
-from pre_commit_sbt.err.exceptions import FailedCommandError
+from pre_commit_sbt.err.failed_command_error import FailedCommandError
 from pre_commit_sbt.run_sbt_command import main_async
