+++ tests/unit/lsp/receive_test.py
@@ -4,7 +4,7 @@ from asyncio import StreamReader
 import pytest
 
 from pre_commit_sbt.err.error_msgs import COMMAND_FAILED
-from pre_commit_sbt.err.exceptions import FailedCommandError
+from pre_commit_sbt.err.failed_command_error import FailedCommandError
 from pre_commit_sbt.lsp.receive import read_until_complete_message
